+++ CHANGELOG.md
@@ -22,3 +22,4 @@ Formato de CHANGELOG baseado em [Keep a Changelog](https://keepachangelog.com/pt
 - Componente `Step Bar`
 - Componente `Table`
 - Componente `Tag`
+- Componente `Popover`